% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_est_expand.R
\name{baseline_est_expand}
\alias{baseline_est_expand}
\title{For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}}
\usage{
baseline_est_expand(
  data_train,
  data_predict,
  n_sim = 1000,
  response,
  fixef,
  ranef = NULL,
  offset
)
}
\arguments{
\item{data_train}{Data to train on.}

\item{data_predict}{Data to predict on}

\item{n_sim}{number of simulations to preform. Default setting is n_sim = 1000}

\item{fixef}{Fixed effekts}

\item{ranef}{Random effekts, default is NULL}

\item{offset}{Offset, can be NULL}
}
\value{
data
}
\description{
For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}
}
\examples{
\dontrun{
data_aggregated <- data.table::as.data.table(data_fake_nowcasting_county_aggregated)
n_sim <- 10
data_train <- data_aggregated[cut_doe< "2019-06-30"]
data_predict <- data_aggregated
response <- "n_death"
fixef <- "1 + sin(2 * pi * (week) / 53) + cos(2 * pi * (week ) / 53) + year"
ranef <- "(1|location_code)"
offset <- "log(pop)"
base<- baseline_est_expand(data_train, data_predict, n_sim, response, fixef, ranef, offset)
}
}
