% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_est_simple.R
\name{baseline_est_simple}
\alias{baseline_est_simple}
\title{For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}}
\usage{
baseline_est_simple(data_train, data_predict, n_sim = 1000, formula, offsett)
}
\arguments{
\item{data_train}{Data to train on.}

\item{data_predict}{Data to predict on}

\item{n_sim}{number of simulations to preform. Default setting is n_sim = 1000}

\item{formula}{Formula to model}

\item{offsett}{True or False depending on whetehr there is an offset in the formula or not.}
}
\value{
Residualplots for all ncor_i and some evaluationmetrixs for each of them as well as a plot containing credible intervals using the simulation
}
\description{
For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}
}
