% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_data.R
\docType{data}
\name{data_fake_attrib_county}
\alias{data_fake_attrib_county}
\title{Fake data for attributable mortality in Norway (counties)}
\format{
\describe{
\item{location_code}{Location code of the Norwegian counties.}
\item{isoyear}{Isoyear.}
\item{isoweek}{Isoweek.}
\item{isoyearweek}{Isoyear and isoweek.}
\item{season}{Season used for influenza like illnesses.}
\item{seasonweek}{Number of weeks from the start of the season.}
\item{pop_jan1_n}{Population size.}
\item{ili_isoweekmean0_6_pr100}{Per hundred ILI, percentage of consultations diagnosed as influenza like illnesses.}
\item{ili_isoweekmean7_13_pr100}{ili_isoweekmean0_6_pr100 lagged by one week.}
\item{heatwavedays_n}{Number of days of the week that were considered a heatwave}
\item{deaths_n}{Number of deaths}
}
}
\usage{
data_fake_attrib_county
}
\description{
Fake data for attributable mortality in Norway (counties)
}
\keyword{datasets}
