% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_death_data.R
\docType{data}
\name{data_fake_death}
\alias{data_fake_death}
\title{Fake data for mortalityregistration}
\format{
\describe{
\item{DoE}{Date og event}
\item{DoR}{Date of registration}
}
}
\usage{
data_fake_death
}
\description{
Fake data for mortalityregistration
}
\keyword{datasets}
