% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_clean.R
\docType{data}
\name{data_fake_death_clean}
\alias{data_fake_death_clean}
\title{Cleaned fake data for mortalityregistration}
\format{
\describe{
\item{DoE}{Date og event}
\item{DoR}{Date of registration}
}
}
\usage{
data_fake_death_clean
}
\description{
Cleaned fake data for mortalityregistration
}
\keyword{datasets}
