% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_attrib_data.R
\docType{data}
\name{data_fake_norway}
\alias{data_fake_norway}
\title{Fake data for mortality in Norway nationaly}
\format{
\describe{
\item{location_code}{Location code}
\item{week}{Week}
\item{season}{Season used for influenza like ilnesses}
\item{yrwk}{Year and week}
\item{x}{Number of weeks from the start of the season}
\item{pop}{Population size}
\item{pr100_ili}{Pr hundered ILI, precentage og consultations diagnised as influenza like illnesses}
\item{pr100_ili_lag_1}{pr100_ili_lag_1}
\item{temperature}{ temperature}
\item{temperature_high}{temperature_high}
\item{deaths}{deaths}
}
}
\usage{
data_fake_norway
}
\description{
This data is licensed under specific conditions noted in the details section.
}
\details{
The permission to use the data is granted on condition that:
\itemize{
\item The data will not be used for commercial purposes
\item The source will be acknowledged. A copyright notice will have to be visible on any printed or electronic publication using this dataset
}
}
\keyword{datasets}
