% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_data.R
\docType{data}
\name{data_fake_nowcasting_nation_raw}
\alias{data_fake_nowcasting_nation_raw}
\title{Fake data for mortality registration}
\format{
\describe{
\item{doe}{Date of event}
\item{dor}{Date of registration}
}
}
\usage{
data_fake_nowcasting_nation_raw
}
\description{
Fake data for mortality registration
}
\keyword{datasets}
