% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_attrib.R
\name{est_attrib}
\alias{est_attrib}
\title{Estimates attributable mortality}
\usage{
est_attrib(fit, data_observed, data_reference)
}
\arguments{
\item{fit}{A model fit}

\item{data_observed}{The observed data}

\item{data_reference}{The data with reference values}
}
\description{
Estimates attributable mortality
}
