% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{fit_attrib}
\alias{fit_attrib}
\title{fit_attrib}
\usage{
fit_attrib(
  data,
  outcome = "deaths",
  exposures = list(influenza = "linear", temperature = "linear")
)
}
\arguments{
\item{data}{x}

\item{outcome}{x}

\item{exposures}{a}
}
\description{
fit_attrib
}
