% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_attrib_data.R
\name{gen_fake_attrib_data}
\alias{gen_fake_attrib_data}
\title{Generates fake attributable data}
\usage{
gen_fake_attrib_data(n_locations = 11)
}
\arguments{
\item{n_locations}{Telling how many locations one watns in the output data}
}
\description{
This function generates one dataset
}
