% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_correction_fn_expanded}
\alias{nowcast_correction_fn_expanded}
\title{For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}}
\usage{
nowcast_correction_fn_expanded(data, n_week_adjusting, offset)
}
\arguments{
\item{data}{Data generated with nowcast_aggregate containing the part of the dataset that the model should train on.}

\item{n_week_adjusting}{Number of weeks to correct}

\item{offset}{Boolian value which is set to true if offset(log(pop)) is a part of the formula}
}
\value{
nowcast_correction_object including corrected data for all weeks in n_wwk_adjust and the model fits for all weeks
}
\description{
For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}
}
\examples{
data <- attrib::data_fake_nowcasting_aggregated
n_week_adjusting <- 8
n_week_train <- 52
n_week_start <- n_week_adjusting + n_week_train
date_0 <- data[nrow(data),]$cut_doe #last date in the dataset, assume the dataset is ordered.
data <- data[cut_doe >= (date_0 - n_week_start*7 + 1), ]
nowcast_correction_object <- nowcast_correction_fn_expanded(data, n_week_adjusting, offset = TRUE )
}
