% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_correction_sim}
\alias{nowcast_correction_sim}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
nowcast_correction_sim(nowcast_correction_object, n_sim = 500)
}
\arguments{
\item{nowcast_correction_object}{object returned from function nowcast_correction_fn_expanded}

\item{n_sim}{Number of simulations}
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
