% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_eval.R
\name{nowcast_eval}
\alias{nowcast_eval}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
nowcast_eval(nowcast_object, n_week_adjusting)
}
\arguments{
\item{nowcast_object}{Object from the function nowcast}

\item{n_week_adjusting}{Number of weeks to adjust}
}
\value{
Residualplots for all ncor_i and some evaluationmetrixs for each of them as well as a plot containing credible intervals using the simulations
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\examples{

data <- attrib::data_fake_nowcasting_aggregated
n_week_adjusting <- 8
data_correct <- nowcast_eval(data, n_week_adjusting )
}
