% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_expand.R
\name{nowcast_exp}
\alias{nowcast_exp}
\title{For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}}
\usage{
nowcast_exp(
  data_aggregated,
  offset,
  n_week_adjusting,
  n_week_training,
  nowcast_correction_fn = nowcast_correction_fn_negbin_mm,
  nowcast_correction_sim_fn = nowcast_correction_sim_neg_bin
)
}
\arguments{
\item{data_aggregated}{Aggregated dataset from the function npowcast_aggregate}

\item{offset}{offset}

\item{n_week_adjusting}{Number of weeks to correct}

\item{n_week_training}{Number of weeks to train on}

\item{nowcast_correction_fn}{Correction function. The deafault is nowcast_correction_fn_expanded. Must return the same as this function.}

\item{nowcast_correction_sim_fn}{Simmulatoin function. Must return a datatable with the following collumns  "n_death", "sim_value", "cut_doe", "ncor" and simmulations for equally many weeks as n_week_adjust.}
}
\value{
Dataset including the corrected values for n_death
}
\description{
For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}
}
\examples{

data <- attrib::data_fake_nowcasting_aggregated
n_week_adjusting <- 8
n_week_training <- 12
nowcast_object <- nowcast_exp(data, n_week_adjusting,n_week_training , offset = TRUE)
}
