% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Generates simulations of expected mortalities by simulating the model coefficiants.}
\usage{
sim(fit, data)
}
\arguments{
\item{fit}{A model fit created by fit_attrib}

\item{data}{The data with eather observed values or referance values.}
}
\value{
A dataset with 500 simulation og the expected response for each row in the orriginal dataset.
}
\description{
With the given fit from fit_attrib the funcion sim, from package arm, is used to generate 500 simulations
of all the coefiecients, from there respective posterior distributions.
This is then used to compute the expected response for all simulations and rows in the input dataset.
}
\details{
\code{vignette("intro", package="attrib")}
}
\examples{

response <- "deaths"
fixef <- "pr100_ili_lag_1 + sin(2 * pi * (week - 1) / 52) + cos(2 * pi * (week - 1) / 52)"
ranef <- " (pr100_ili_lag_1| season)"
offset <- "log(pop)"

data = attrib::data_fake_nation

fit <- fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)

sim(fit, data)
}
