% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_est.R
\name{baseline_est}
\alias{baseline_est}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
baseline_est(data_train, data_predict, n_sim = 1000, formula, offsett)
}
\arguments{
\item{data_train}{Data to train on.}

\item{data_predict}{Data to predict on}

\item{n_sim}{number of simulations to preform. Default setting is n_sim = 1000}

\item{formula}{Formula to model}

\item{offset}{True or False depending on whetehr there is an offset in the formula or not.}
}
\value{
Residualplots for all ncor_i and some evaluationmetrixs for each of them as well as a plot containing credible intervals using the simulation
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\examples{

 data_aggregated <- data.table::as.data.table(attrib::data_fake_nowcasting_aggregated)
 n_week_training <- 50
 n_week_adjusting <- 8
 nowcast_object <- nowcast(data_aggregated= data_aggregated,
   n_week_training = 50, n_week_adjusting = 8)
 nowcast_eval_object <- nowcast_eval(nowcast_object, n_week_adjusting)
}
