% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_attrib.R
\name{est_attrib}
\alias{est_attrib}
\title{Estimates attributable mortality and irr}
\usage{
est_attrib(data, exposures, response)
}
\arguments{
\item{data}{The observed data}

\item{exposures}{The exposures that will be given attributable deaths, must exict as exp_mort + exposure in the data}

\item{response}{The name of the response column}
}
\description{
Estimates attributable mortality and irr
}
