% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_attrib.R
\name{est_attrib}
\alias{est_attrib}
\title{Estimates attributable mortality and irr}
\usage{
est_attrib(data)
}
\arguments{
\item{data}{The output data from est_mort on long format with all attr in one colum with there values in clumn "value"}
}
\description{
Estimates attributable mortality and irr
}
