% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_attrib.R
\name{est_attrib}
\alias{est_attrib}
\title{Estimates attributable mortality}
\usage{
est_attrib(fit, data, exposures)
}
\arguments{
\item{fit}{A model fit}

\item{data}{The observed data}

\item{exposures}{The exposures that will be given attributable deaths}
}
\description{
Estimates attributable mortality
}
