% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mort.R
\name{est_mort}
\alias{est_mort}
\title{Estimates expected mortality}
\usage{
est_mort(fit, data, exposures, response)
}
\arguments{
\item{fit}{A model fit}

\item{data}{The observed data}

\item{exposures}{The exposures that will get reference expected mortalities}

\item{response}{The name of the response column}
}
\description{
Estimates expected mortality
}
