% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{fit_attrib}
\alias{fit_attrib}
\title{Data fit}
\usage{
fit_attrib(
  data,
  response,
  fixef,
  ranef,
  dist_family = "poisson",
  offset = NULL
)
}
\arguments{
\item{data}{The observed data to be fitted.}

\item{response}{The response}

\item{fixef}{The fixed effects}

\item{ranef}{The random effects}

\item{dist_family}{Family}

\item{offset}{The offsets.}
}
\value{
The model fit of the data with additional attributes offset, response and fit_fix.
Offset and response are the same as in the input and fit_fix is the linear model of the fix effects.

For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\description{
Data fit using glmer from lme4 with family poisson to fit the dataset with the given formula.
}
\examples{

response <- "deaths_n"

fixef <- "ili_isoweekmean7_13_pr100 +
  sin(2 * pi * (isoweek - 1) / 52) +
  cos(2 * pi * (isoweek - 1) / 52)"
ranef <- " (ili_isoweekmean7_13_pr100| season)"
offset <- "log(pop_jan1_n)"

data <- attrib::data_fake_attrib_nation

fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)
}
