% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{fit_attrib}
\alias{fit_attrib}
\title{fit_attrib}
\usage{
fit_attrib(data, response, fixef, ranef, offset = NULL)
}
\arguments{
\item{data}{x}

\item{response}{x}

\item{fixef}{x}

\item{ranef}{x}

\item{offset}{x}
}
\description{
fit_attrib
}
