% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{fit_attrib}
\alias{fit_attrib}
\title{fit_attrib}
\usage{
fit_attrib(dates = FluMoDL::greece$daily$date,
  outcome = FluMoDL::greece$daily$deaths, exposure_values = list(tg =
  FluMoDL::greece$daily$temp), exposure_types = list(tg = "cubic"))
}
\arguments{
\item{dates}{x}

\item{outcome}{x}

\item{exposure_values}{x}

\item{exposure_types}{a}
}
\description{
fit_attrib
}
