% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{fit_attrib}
\alias{fit_attrib}
\title{Data fits}
\usage{
fit_attrib(data, response, fixef, ranef, offset = NULL)
}
\arguments{
\item{data}{The observed data to be fitted.}

\item{response}{The response}

\item{fixef}{The fixed effects}

\item{ranef}{The random effects}

\item{offset}{The offsets.}
}
\value{
The model fit of the data with additional attributes offset and fit_fix.
Offset is the same offset as in the input and fit_fix is the linear model of the fix effects.

For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\description{
Data fit using glmer from lme4 and family poisson to fit the dataset with the given formula.
}
\examples{

response = "deaths"
fixef = "pr100_ili + sin(2 * pi * (week - 1) / 52) + cos(2 * pi * (week - 1) / 52)"
ranef = " (pr100_ili| season)"
offset <- "log(pop)"

data = data.table::data.table(deaths = c(100, 110, 150, 160), week = c(10, 11, 10, 11),
                  season = c("2018/2019", "2018/2019", "2019/2020", "2019/2020"),
                  pr100_ili = c(0.2, 0.21, 0.35, 0.36),
                  pop = c(50000, 50000, 50000, 50000))

fit_attrib(data = data, response = response, fixef = fixef, ranef = ranef, offset = offset)

}
