% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attrib.R
\name{get_attrib}
\alias{get_attrib}
\title{get_attrib}
\usage{
get_attrib(attrib, use_blup = FALSE, tag, range, sub = NULL)
}
\arguments{
\item{attrib}{x}

\item{use_blup}{x}

\item{tag}{x}

\item{range}{a}

\item{sub}{a}
}
\description{
get_attrib
}
