% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
nowcast()
}
\arguments{
\item{data_clean}{Cleaned dataset from the function npowcast_clean}

\item{n_week}{Number of weeks to correct}
}
\value{
Dataset including the corrected values for n_death
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\examples{
\dontrun{

data <- attrib::data_fake_death_clean
n_week <- 8

data_correct <- nowcast(data, n_week)
}
}
