% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
nowcast(
  data_aggregated,
  n_week_adjusting,
  n_week_training,
  nowcast_correction_fn = nowcast_correction_fn_expanded
)
}
\arguments{
\item{data_aggregated}{Aggregated dataset from the function npowcast_aggregate}

\item{n_week_adjusting}{Number of weeks to correct}

\item{n_week_training}{Number of weeks to train on}

\item{nowcast_correction_fn}{Correction function. Must return a table with columnames ncor0_i for i in 0:n_week and cut_doe. The default uses "n_death ~ n0_i" for all i in 0:n_week.}
}
\value{
Dataset including the corrected values for n_death
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
\examples{
\dontrun{

data <- attrib::data_fake_nowcasting_aggregated
n_week_adjusting <- 8
n_week_training <- 12
data_correct <- nowcast(data, n_week_adjusting,n_week_training )
}
}
