% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_correction_fn_simple}
\alias{nowcast_correction_fn_simple}
\title{For more details see the help vignette:
\code{vignette("intro", package="attrib")}}
\usage{
nowcast_correction_fn_simple(data, n_week_adjusting)
}
\arguments{
\item{data}{cleaned data to perform correction formula on}

\item{n_week_adjusting}{Number of weeks to correct}
}
\description{
For more details see the help vignette:
\code{vignette("intro", package="attrib")}
}
