% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_eval.R
\name{nowcast_eval}
\alias{nowcast_eval}
\title{For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}}
\usage{
nowcast_eval(nowcast_object, n_week_adjusting)
}
\arguments{
\item{nowcast_object}{Object from the function nowcast}

\item{n_week_adjusting}{Number of weeks to adjust}
}
\value{
Residualplots for all ncor_i and some evaluationmetrixs for each of them as well as a plot containing credible intervals using the simulations
}
\description{
For more details see the help vignette:
\code{vignette("nowcast", package="attrib")}
}
\examples{

 data_aggregated <- data.table::as.data.table(attrib::data_fake_nowcasting_aggregated)
 n_week_training <- 50
 n_week_adjusting <- 8
 nowcast_object <- nowcast(data_aggregated= data_aggregated,
   n_week_training = 50, n_week_adjusting = 8, offset = TRUE)
 nowcast_eval_object <- nowcast_eval(nowcast_object, n_week_adjusting)
}
